Add autograding results to gradingview (#381)
* Added autograding_results to grading_view

* Updated test

* Moved autogradingResults from :grade to :question

* Applied refactoring suggestion

* Refactored grading_view

* Isolate helper functions into new module
